==> MuscaChasingMusca/01_Read_raw_xyz_csv_trimDataToStartStopNSmoothen.ipynb <==
## this code trims the tracked points to the correct start and stop frames as manually annotated from videos.

#### Tanvi Deora, March 12 2025

import pandas as pd
import numpy as np
import glob
from scipy import signal

import plotly.graph_objects as go

files = glob.glob('../dataFolders/MuscaChasingMusca/Musca-Musca_Chase_ProjectFiles_Metadata/*Trial*xyzpts.csv')
files = np.delete(files, [1])
files

info = pd.read_csv('../dataFolders/MuscaChasingMusca/metadata.csv')
info

def trimPointsToVideo(df, start, stop):
    new_df = df.loc[start:stop, :]
    return(new_df)

def subsample(data, framerate, target_framerate):
    fs_target = target_framerate    # target sampling rate
    if framerate == fs_target:
        data_1k = data
    else:
        factor = int(framerate // fs_target) 
        # Subsample by keeping every nth sample
        data_sub = data.loc[::factor, :]
    return(data_sub)

for f in files:
    df = pd.read_csv(f)
     
    df.head()
    # get the start, top points
    frames = info.loc[info['filename'] == (f.split('\\')[-1]),['start','end']].values.astype('int32')[0]
    # add -1 to adjust to python starting from 0th number
    start, stop = frames[0]-1, frames[1]-1
    new_df = trimPointsToVideo(df, start, stop)
    
    # interpolate to get rid of NaNs
    inter_df = new_df.interpolate(method='spline', order = 3, axis=0, limit = 10)
    
    # get framerate to subsample everything to targer_framerate
    framerate = [row['framerate'] for ind, row in info.iterrows() if row['filename'][:-10] in f.split('\\')[-1]][0]
    df_1k = subsample(inter_df, framerate, target_framerate = 100)
    
    # smoothen data with savgol s filter
    smooth_df = pd.DataFrame(index = df_1k.index, columns = df_1k.columns)
    for name, series in df_1k.items():
        smooth_df[name] = signal.savgol_filter(series.values,
                           6, # window size used for filtering
                           3) # order of fitted polynomial
        
    name = f.split('\\')[-1][:-11]
    smooth_df.to_csv('../dataFolders/MuscaChasingMusca/Outputs/Sampled_100fps/' + name + '_smoothened_interpol_sampled100fps_xyzpts.csv')
