==> MuscaChasingMusca/02_Read_raw_csv_GenerateCOM.ipynb <==
import pandas as pd
import numpy as np
import glob

import plotly.graph_objects as go

def makeFigBodyHeading(df, fig):
    # add head and abdomen dots
    # chasee
    fig.add_trace(go.Scatter3d(x = df.loc[:,'chasee_head_X']
                 , y = df.loc[:, 'chasee_head_Y']
                 , z = df.loc[:, 'chasee_head_Z']
                 , mode = 'markers'
                 , marker = dict(size=2, color="brown")
                ))

#     fig.add_trace(go.Scatter3d(x = df.loc[:,'chasee_abdomen_X']
#                  , y = df.loc[:, 'chasee_abdomen_Y']
#                  , z = df.loc[:, 'chasee_abdomen_Z']
#                  , mode = 'markers'
#                  , marker = dict(size=2, color="brown")
#                 ))
    # chaser
    fig.add_trace(go.Scatter3d(x = df.loc[:,'chaser_head_X']
                 , y = df.loc[:, 'chaser_head_Y']
                 , z = df.loc[:, 'chaser_head_Z']
                 , mode = 'markers'
                 , marker = dict(size=2, color="brown")
                ))

#     fig.add_trace(go.Scatter3d(x = df.loc[:,'chaser_abdomen_X']
#                  , y = df.loc[:, 'chaser_abdomen_Y']
#                  , z = df.loc[:, 'chaser_abdomen_Z']
#                  , mode = 'markers'
#                  , marker = dict(size=2, color="brown")
#                 ))

    # add lines for body
    for frame, data in df.iterrows():
        fig.add_trace(go.Scatter3d(x = [data['chasee_head_X'], data['chasee_abdomen_X']]
                 , y = [data['chasee_head_Y'], data['chasee_abdomen_Y']]
                 , z = [data['chasee_head_Z'], data['chasee_abdomen_Z']]
                 , mode = 'lines'
                 , marker = dict(size=4, color="gray")
                              ))

    for frame, data in df.iterrows():
        fig.add_trace(go.Scatter3d(x = [data['chaser_head_X'], data['chaser_abdomen_X']]
                 , y = [data['chaser_head_Y'], data['chaser_abdomen_Y']]
                 , z = [data['chaser_head_Z'], data['chaser_abdomen_Z']]
                 , mode = 'lines'
                 , marker = dict(size=4, color="gray")
                              ))

def makeFigure_COMtracks(df, fig):
     
    fig.add_trace(go.Scatter3d(x = df.loc[:,'chasee_com_x']
             , y = df.loc[:, 'chasee_com_y']
             , z = df.loc[:, 'chasee_com_z']
             , mode = 'lines'
             , marker = dict(size=4, color="black"), name = 'Chasee'
            ))

    fig.add_trace(go.Scatter3d(x = df.loc[:,'chaser_com_x']
             , y = df.loc[:, 'chaser_com_y']
             , z = df.loc[:, 'chaser_com_z']
             , mode = 'lines'
             , marker = dict(size=4, color="red"), name = 'Chaser'
            ))

def makeFigure_COMtracks_colorcoded(df, fig):
      
    fig.add_trace(go.Scatter3d(x = df.loc[:,'chasee_com_x']
             , y = df.loc[:, 'chasee_com_y']
             , z = df.loc[:, 'chasee_com_z']
             , mode = 'markers'
             , marker = dict(
                 size=4
                 ,color=np.arange(len(df)) #set color equal to a variable
                 , colorscale='YlOrRd', # one of plotly colorscales
                 )
            , name = 'Chasee'
            ))

    fig.add_trace(go.Scatter3d(x = df.loc[:,'chaser_com_x']
             , y = df.loc[:, 'chaser_com_y']
             , z = df.loc[:, 'chaser_com_z']
             , mode = 'markers'
             , marker = dict(
                 size=4
                 ,color=np.arange(len(df)) #set color equal to a variable
                 , colorscale='YlGnBu', # one of plotly colorscales
                 )
            , name = 'Chaser'
            ))

def get_COM(head,abdomen):
    COM = pd.DataFrame(data = None)
    COM['COM_X'] = (head.iloc[:,0] + abdomen.iloc[:,0])/2
    COM['COM_Y'] = (head.iloc[:,1] + abdomen.iloc[:,1])/2
    COM['COM_Z'] = (head.iloc[:,2] + abdomen.iloc[:,2])/2
    return(COM)

files = glob.glob('../dataFolders/MuscaChasingMusca/Outputs/Sampled_100fps/*smoothened*_xyzpts.csv')

files

key = pd.read_csv('../dataFolders/MuscaChasingMusca/ReadMe.txt', header = None, sep='\t')
key.index = key[0]
key.drop(columns = 0)
new_key = key.to_dict()[1]
new_key

info = pd.read_csv('../dataFolders/MuscaChasingMusca/metadata.csv')
info

for f in files:
    df = pd.read_csv(f,index_col=0)
    new_names = []
    for names in df.columns:
        a,b = names.split('_')
        new_name = new_key[a] + '_' + b
    #     print(names, new_name)
        new_names.append(new_name) 

    df.columns = new_names

    # get coordinates for chasee and chaser
    Chasee_head = df.loc[:, ['chasee_head_X', 'chasee_head_Y', 'chasee_head_Z']]
    Chasee_abdomen = df.loc[:, ['chasee_abdomen_X', 'chasee_abdomen_Y', 'chasee_abdomen_Z']]

    Chaser_head = df.loc[:, ['chaser_head_X', 'chaser_head_Y', 'chaser_head_Z']]
    Chaser_abdomen = df.loc[:, ['chaser_abdomen_X', 'chaser_abdomen_Y', 'chaser_abdomen_Z']]

    # get COM for chaser and chasee
    Chasee_COM = get_COM(Chasee_head, Chasee_abdomen)
    Chasee_COM.columns = ['chasee_com_x', 'chasee_com_y', 'chasee_com_z']
    Chaser_COM = get_COM(Chaser_head, Chaser_abdomen)
    Chaser_COM.columns = ['chaser_com_x', 'chaser_com_y', 'chaser_com_z']

    # add those to dataFrame
    df = pd.concat([df, Chasee_COM, Chaser_COM], axis=1)


    name = '_'.join(f.split('\\')[-1].split('_')[:-2])

    fig = go.Figure()
    makeFigure_COMtracks(df, fig)
    makeFigBodyHeading(df, fig)
#     fig.show()
    fig.write_html('../dataFolders/MuscaChasingMusca/Sampled100fps_Figures/Trajectories/' + name + '_COM_trajectory.html')
    
    fig1 = go.Figure()
    makeFigure_COMtracks_colorcoded(df, fig1)
#     fig.show()
    fig1.write_html('../dataFolders/MuscaChasingMusca/Sampled100fps_Figures/Trajectories/' + name + '_COM_trajectory_colorCoded.html')
    
    df.to_csv('../dataFolders/MuscaChasingMusca/Outputs/Sampled_100fps/COM/' + name + '_COM.csv')
